requirements.txt

Flask==2.2.3
requests==2.28.2
msal==1.21.0
python-dotenv==1.0.0
pytz==2023.2
